requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.6
